refactor: decouple pandas postprocessing operator (#18710)
